packages/opencloning-db/src/pyproject.toml
[tool.poetry]
authors = ["Manuel Lera-Ramirez <manulera14@gmail.com>"]
description = "Database for OpenCloning, a web application to generate molecular cloning strategies in json format, and share them with others."
license = "MIT"
name = "opencloning-db"
version = "0.1.0"
package-mode = false
readme = "README.md"
repository = "https://github.com/manulera/OpenCloning_db"

[tool.poetry.dependencies]
python = "^3.11"
opencloning = {git = "https://github.com/manulera/OpenCloning_backend"}
sqlalchemy = "^2.0.41"
fastapi-pagination = "^0.15.10"
PyJWT = "^2.10.1"
pwdlib = {version = "^0.2.1", extras = ["argon2"]}
pydna = {git = "https://github.com/pydna-group/pydna", rev = "25772b76ac7dfff52b9f8bec4cc6e0279372afc0"}
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
flake8-bugbear = "^24.2.6"
black = "^24.2.0"
pre-commit = "^3.6.2"
watchfiles = "^0.21.0"
nbstripout = "^0.8.1"

[tool.poetry.group.test.dependencies]
pytest = "8.3.4"
pre-commit = "^3.6.2"
pytest-cov = "^4.1.0"
pytest-rerunfailures = "^14.0"
respx = "^0.21.1"


[tool.poetry.group.ipython.dependencies]
ipython = "^8.20.0"
ipykernel = "^6.28.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
skip-string-normalization = true
line-length = 119
